crossword/requirements.txt
pillow
pytest
hypothesis